modify setup
