+++ .env.sample
@@ -1,3 +1,10 @@
-DATABASE_URL= //your_database_url_here
-PORT= //Your_preferred_port_number_here
-SECRET_KEY= \\your_secret_key_here
+
+FLASK_APP=app
+FLASK_DEBUG=True
+FLASK_RUN_PORT=8000
+FLASK_RUN_HOST=0.0.0.0
+
+SQLALCHEMY_DATABASE_URI=postgresql://username:password@localhost:5432/postgresql
+
+JWT_SECRET_KEY=your_secret_key_here
+JWT_ACCESS_TOKEN_EXPIRES=7200
