+++ docker-compose.yml
@@ -1,0 +1,33 @@
+
+version: "3.9"
+
+services:
+  flask_app:
+    container_name: book_app
+    image: book_api
+    build: .
+    ports:
+      - "4000:4000"
+    environment:
+      - DATABASE_URL=postgresql://username:password@flask_db:5432/postgres
+      - FLASK_ENV=development  # Set to 'production' in a production environment
+    depends_on:
+      - flask_db
+  flask_db:
+    container_name: flask_db
+    image: postgres:12
+    ports:
+      - "5432:5432"
+    environment:
+      - POSTGRES_HOST_AUTH_METHOD=trust
+      - POSTGRES_USER=postgres
+      - POSTGRES_DB=postgres
+    volumes:
+      - pgdata:/var/lib/postgresql/data
+
+volumes:
+  pgdata: {}
+
+
+
+
