+++ extensions.py
@@ -1,0 +1,9 @@
+from flask_cors import CORS
+from flask_jwt_extended import JWTManager
+from flask_migrate import Migrate
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+migrate = Migrate()
+cors = CORS()
+jwt = JWTManager()
