--- todos
@@ -1,1 +1,0 @@
-1. Add docker setup
